adds image paste from clipboard (upload to github)
will check the clipboard if a image was pasted, if not it will do the
default paste operartion (paste the text). But if a image is found it
will ask the user for the path to paste it to, the path is relative to
the repository's root directory.
So for example ur github repository lookes like this:

/  (Repository Root)
|
|--img
|  |--myImage.png
|  |--AnotherImage.png
|  '...
|
'--md_files
   |--somefile.md
   '...

and u want to upload a image to the img folder, you would enter
"img/myUpload.png" and myUpload.png would be added to the img folder.
Be sure to also enter the image type. If u supply a path to an already
existing image it will be overwritten.