adds image upload support from clipboard (upload to github)
on paste event (strg-v) the clipboard is checked for a image, if one is
found the user will be prompt for the image name with a pre entered name
(img_[unixtimestamp].png). the user can then enter either a name, accept
the given name, or abort (in which case nothing will happen).
image folder and branch are asked at github publish settings, commit
message can be configured in settings (below github commit message
settings), default commit message is 'Image upload.'
a missing / on the end of the given folder will be added if not given.

after a upload to github the raw.githubusercontent.com link will be
inserted. for example:
![](https://raw.githubusercontent.com/username/repo/imgBranch/imgFolder/imgName)

if the given image branch is gh-pages the github.io link will be
inserted instead. for example:
![](http://username.github.io/repo/imageFolder/imgName)