start stackedit with enabled githubPublisher
adds the opportunity to start stackedit with a injected publisher
(+downloader) so that it starts and loads a given file. This is enabled
by appending json formated text to the link. here is a example:

http://stackedit.io/editor#{"type":"github","publishAttributes":{"username": ... }}

example json string after the #:
{
        "type"              : "github",
        "publishAttributes" :
        {
                "username"   : "pete",
                "title"      : "myAwesomeMarkdownFile.md",
                "repository" : "myDocs",
                "branch"     : "master",
                "format"     : "markdown",
                "path"       : "path/to/my/text.md"
        }
}

"type" is the given provider type u want to access, someone could add
other providers if they like. if type "github" is given a fileDesc with
the given "publishAttributes" will be created and the file content will
be directly downloaded from github.
if the user has no github token (not logged in) he will be redirected to
the github auth page on editor start, and the download will proceed
after login.

if a file with the same title is already present inside stackedit, it's
content will be replaced with the content of the github file,
publishAttributes will also be overwritten (repo, branch, path, etc). so
be shure to not open a file where u have unpublished changes, because
these will be lost.

if a file which does not exists is given, an empty file will be opened, which
can be published to the given location.

- username   : github repository username (can differ from your username)
- title      : the file Title
- repository : github repository
- branch     : github branch
- format     : file format
- path       : github file path
- imgPath    : folder for image uploads (optional, see pull request #620)
- imgBranch  : branch for image uploads (optional, see pull request #620)