stackedit auto-download given file on startup
adds the opportunity to start stackedit with a url and title
so that it starts and loads a given file. This is enabled by appending
json formated text to the link. here is a example:

stackedit.io/editor#{"type":"download","importParameters":{"url": ... }}

json string after the #:
{
        "type"             : "download",
        "importParameters" :
        {
                "url"   : "http://url/to/my/file.md",
                "title" : "myAwesomeFile.md",
        }
}

"type" is the given provider type u want to use, someone could add
other providers if they like. if type "download" is given a fileDesc with
the given "importParameters" will be created and the file content will
be directly downloaded from the given url.
the title will be checked against all other loaded documents, if a file
with the same title already exists it's content will be overwritten by
the downloaded content.

- url   : the url where the file content can be downloaded
- title : the title which will be used