Removes 'newline' configuration.
Pagedown-extra newline config will add a <br> each time it detects a single newline.
This is annoying because it encourages our users of the editor to preserve long lines,

which again makes it harder to diff changes

NOTE: In my opinion, it's more faithful to markdown to remove this setting since
it works that way for standard markdown. In my opinion, it's the most reasonable default.